backend/requirements.txt
# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
redis>=5.0.1

# Data Processing
pandas>=1.5.0
numpy>=1.21.0
pyspark>=3.3.0  # Optional, for large-scale data processing
yfinance>=0.2.0

# Configuration Management
pyyaml>=6.0.0

# Data Storage
pyarrow>=10.0.0  # Parquet support

# Testing
pytest>=7.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0
pytest-mock>=3.12.0
pytest-timeout>=2.2.0
pytest-xdist>=3.5.0

# Development
black>=24.0.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0

# Visualization
matplotlib>=3.5.0
seaborn 
openpyxl

# Trading System
coloredlogs>=15.0
scikit-learn>=1.0.0
ta>=0.10.0  # Technical analysis library
requests>=2.28.0
python-dateutil>=2.8.0

# API
flask>=2.0.0
flask-cors>=4.0.0  # For cross-origin requests
gunicorn>=20.1.0  # For production deployments